Optimize CPU emulation performance
- Use Uint8Array instead of Uint16Array for opcode dispatch table (saves 4KB)
- Inline opcode lookup to eliminate function call overhead in hot path
- Add fast-path for RAM access in get_mem/set_mem (most common case)
- Reorganize memory access branches for better CPU branch prediction

These optimizations target the main bottlenecks in the emulation loop:
- Reduced memory footprint improves cache utilization
- Inlined lookups eliminate function call overhead
- Fast-path returns reduce average branches per memory access

+++ utils/rle-array.js
@@ -1,7 +1,7 @@
 export class RLEArray {
   constructor(maxCapacity) {
     this._length = 0;
-    this._indexTable = new Uint16Array(maxCapacity);
+    this._indexTable = new Uint8Array(maxCapacity);
     this._values = [];
   }
